requirements.txt
requests>=2.31.0
dropbox>=11.36.0
python-dotenv>=1.0.0
boto3>=1.28.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
